Improves quality of TraceContextOrSamplingFlags (#1185)
This deprecates methods only used in tests, or those that can lead to
null pointers. It backfills JavaDoc and complete test coverage.

Since `TraceContextOrSamplingFlags` primarily called "extracted", this
improves the `toString` based on experience of what's important to know.

Ex. we want to know what we extracted, if any sampling status exists,
and if any extra propagated state exists.