Renames internal ExtraHandler back to ExtraFactory and clarifies (#1190)
When integrating secondary sampling, I found the new name (I caused)
more distracting then helpful. The defensive api which led to renaming
was a cure worse than the disease of a preventable programming error.

Instead, this improves Javadoc, notably mentioning how to use the
factory properly (and not crashing on mistake).