Makes MutableSpan easier with tags() and annotations() method
This makes `MutableSpan` easier, especially in tests, by adding methods
for tags and annotations similar to `zipkin2.Span`. Unlike
`zipkin2.Span`, this presents annotations as a collection of `Map.Entry`.

Doing so keeps us from having to define a new type mostly used in tests.
The allocation-free consumer functions are left in place. This replaces
the 'count' methods formerly there (in order to try to keep method count
down).

One unit test was converted in the process. We can convert the rest of
them following this PR.

Under the scenes, this uses the same `UnsafeArrayMap` we use in baggage.
To port to this model, we had to switch this to use copy-on-write arrays
instead of array list. This will actually allocate less except when a
tag is overwritten as before overwriting didn't allocate an array.
However, this also cheapens usage by a couple copies anyway so it should
be a wash or better for most folks.