Makes MutableSpan easier with tags() and annotations() method (#1196)
This makes `MutableSpan` easier, especially in tests, by adding methods
for tags and annotations similar to `zipkin2.Span`. Unlike
`zipkin2.Span`, this presents annotations as a collection of `Map.Entry`.

Doing so keeps us from having to define a new type mostly used in tests.
The allocation-free consumer functions are left in place. This replaces
the 'count' methods formerly there (in order to try to keep method count
down).