Makes default propagation for non-request spans B3 single
As described in the referenced issue. In-process messaging sometimes
uses propagation, and the default of B3 multi makes the overhead quite
bad. This switches to B3 single which significantly redudes that burden.

https://github.com/spring-cloud/spring-cloud-sleuth/pull/1607
^^ already changed this in sleuth. this is just defaulting here

We should note that B3 single is widely supported at this point, and all
our instrumentation extend Request when injecting.

See #1020 as this was the change that allowed requests to differentiate