Attempts to deflake servlet after Jetty 9 update
After moving to Jetty 9, we get occasional flakes seeing errors in servlet
requests. This moves the atomic boolean around to help rule out redundnant calls
as a problem. Possibly, though it could be that something is eating the error
or that the request is completing prior to the error being visible somehow.