Adds RpcTracing.propagation() and fixes grpc-trace-bin propagation
When gRPC dropped census, it broke our integration tests. We removed
them, and in doing so lost coverage of the deprecated `grpc-trace-bin`
format, which drifted and broke.

@dimi-nk's company are still using that format. The cleanest way to
support it is to allow overriding of `RpcTracing.propagation()`. This
allows us the most flexibility including backporting the feature
without breaking API.

While I've done this only for RpcTracing, it might make sense for
HTTP and Messaging, as some library specific formats might exist
elsewhere. If no one asks for that, I'll leave it to just RpcTracing.

cc @anuraaga @jeqo @jorgheymans @jcchavezs