Capture row counts for plain statements and batch
Previously, only row counts for executeUpdate on prepared statements was
included as sql.affected_rows. With this change, row counts are also
included for executeUpdate on plain statements, and for executeBatch on
prepared and plain statements.

Capturing row counts on plain execute is a lot trickier, since it´s not
safe to call getUpdateCount behind the application´s back, as it is not
idempotent - we´d unexpectedly change state.