Fixed the httpclient5 licenses.
The original PR that introduced these classes came from 2020, but the squash merge updated their git datetime to 2023, which makes the license formatting check fail on main.

With this change, the headers are now correctly formatted according to the new git datetime, and a master branch build should hopefully succeed.