build: fixes bytecode level and updates build versions (#236) (#1393)
This fixes a problem where we weren't releasing jars with the right
bytecode level, due to a constant release=8. As we no longer can rely on
RetroLambda to rewrite code, I've had to port things to the appropriate
bytecode level manually. I also updated build deps that could be.

Special note: Projects that optionally use types above their bytecode
level *unset* maven.compiler.release in the release profile. This allows
the project to compile classes not at the source level. This is used in
brave core and okhttp, for example.

Signed-off-by: Adrian Cole <adrian@tetrate.io>