ci: deploys bom separately to allow re-do on partial failure (#1406)
Before, the bom project deployed only after the parent succeeded. This
meant a timeout would prevent it from being deployed, even if the
timeout was not critical. This deploys the bom independently so that
such a situation doesn't cause us to have to re-cut a version.

This also fixes where we didn't actually use the release args in tests.

Signed-off-by: Adrian Cole <adrian@tetrate.io>