Consolidates notes on "extra" data
Before this change, notes about state, particularly the "one instance
per plugin" rule were not consolidated on the method that can cause
problems `addExtra`. This migrates notes together with additional
warnings that when we talk about this being used for propagation
plugins, we also mean not for ad-hoc usage, as that can lead to
excessive overhead.

See #1421

Signed-off-by: Adrian Cole <adrian@tetrate.io>