jdbi3: converts to StatementLogger and follows existing style (#1464)
This converts jdbi3 instrumentation to be more like MongoDB in
implementation: configured by instance, not static like mysql connection
driver. It also pares back the data policy to be the same as MySQL: this
allows us a base case to move forward with. Finally, it switches
implementation to a StatementLogger, so that exceptions are visible.

Infrastructure wise, this re-uses the more recent style from
instrumentation here, including docker for the IT.

This is a follow-up to #1460 before we release a version, so API breaks
are not an issue, yet.

Thanks to @reftel for the initial work on this, and @reta on the review
of that. I did my best in RATIONALE.md here to explain why the changes
are important.

---------

Signed-off-by: Adrian Cole <adrian.cole@elastic.co>